requirements.txt
fastapi
uvicorn
requests
PyGithub
python-dotenv
pydantic